requirements.txt
# Core Web Framework
Flask==3.0.0
flask-cors==4.0.1
gunicorn==21.2.0

# OpenAI Integration
openai==1.51.0

# Environment Variables
python-dotenv==1.0.1

# AI/ML Libraries
sentence-transformers==3.1.1
chromadb==0.5.7
numpy==1.26.4

# Utilities
email-validator==2.2.0
requests==2.31.0
